+++ pkg/apis/eunomia/v1alpha1/gitopsconfig_types.go
@@ -19,13 +19,13 @@ type GitConfig struct {
 	SecretRef  string `json:"secretRef,omitempty"`
 }
 
-// GitOpsTrigger represents a trigge, possible type values are change, periodic, webhook.
+// GitOpsTrigger represents a trigger, possible type values are change, periodic, webhook.
 // If token is used the object must be labeled with the following label: "gitops_config.eunomia.kohls.io/webhook_token: <token>"
 type GitOpsTrigger struct {
 	// Type supported types are Change, Periodic, Webhook
 	// +kubebuilder:validation:Enum=Change,Periodic,Webhook
 	Type string `json:"type,omitempty"`
-	// creon expression only valid with the Periodic type
+	// cron expression only valid with the Periodic type
 	Cron string `json:"cron,omitempty"`
 	// webhook secret only valid with webhook type
 	Secret string `json:"secret,omitempty"`
@@ -35,7 +35,7 @@ type GitOpsTrigger struct {
 // +k8s:openapi-gen=true
 type GitOpsConfigSpec struct {
 	// INSERT ADDITIONAL SPEC FIELDS - desired state of cluster
-	// Important: Run "operator-sdk generate k8s" to regenerate code after modifying this file
+	// Important: Run "make generate" to regenerate code after modifying this file
 	// Add custom validation using kubebuilder tags: https://book.kubebuilder.io/beyond_basics/generating_crd.html
 
 	// TemplateSource is the location of the templated resources
@@ -48,20 +48,28 @@ type GitOpsConfigSpec struct {
 	ServiceAccountRef string `json:"serviceAccountRef,omitempty"`
 	// TemplateEngine, the gitops operator config map contains the list of available template engines, the value used here must exist in that list. Identity (i.e. no resource processing) is the default
 	TemplateProcessorImage string `json:"templateProcessorImage,omitempty"`
-	// ResourceHandlingMode represents how resource creation/update should be handled. Supported values are CreateOrMerge,CreateOrUpdate,Patch,None. Default is CreateOrMerge.
-	// +kubebuilder:validation:Enum=CreateOrMerge,CreateOrUpdate,Patch,None
+	// ResourceHandlingMode represents how resource creation/update should be handled. Supported values are Apply,Create,Delete,Patch,Replace,None. Default is Apply.
+	// +kubebuilder:validation:Enum=Apply,Create,Delete,Patch,Replace,None
 	ResourceHandlingMode string `json:"resourceHandlingMode,omitempty"`
 	// ResourceDeletionMode represents how resource deletion should be handled. Supported values are Retain,Delete,None. Default is Delete
 	// +kubebuilder:validation:Enum=Retain,Delete,None
 	ResourceDeletionMode string `json:"resourceDeletionMode,omitempty"`
+	// TemplateProcessorArgs references to the run time parameters, we can pass additional arguments/flags to the template processor.
+	TemplateProcessorArgs string `json:"templateProcessorArgs,omitempty"`
 }
 
 // GitOpsConfigStatus defines the observed state of GitOpsConfig
 // +k8s:openapi-gen=true
 type GitOpsConfigStatus struct {
 	// INSERT ADDITIONAL STATUS FIELD - define observed state of cluster
-	// Important: Run "operator-sdk generate k8s" to regenerate code after modifying this file
+	// Important: Run "make generate" to regenerate code after modifying this file
 	// Add custom validation using kubebuilder tags: https://book.kubebuilder.io/beyond_basics/generating_crd.html
+
+	State            string       `json:"state,omitempty"`
+	StartTime        *metav1.Time `json:"startTime,omitempty"`
+	CompletionTime   *metav1.Time `json:"completionTime,omitempty"`
+	Message          string       `json:"message,omitempty"`
+	LastScheduleTime *metav1.Time `json:"lastScheduleTime,omitempty"`
 }
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object