+++ pkg/util/util.go
@@ -41,11 +41,11 @@ type JobMergeData struct {
 	Action string `json:"action,omitempty"`
 }
 
-// InitializeTemplates initializes the temolates needed by this controller, it must be called at controller boot time
-func InitializeTemplates(jobTempateFileName string, cronJobTemplateFilename string) error {
-	text, err := ioutil.ReadFile(jobTempateFileName)
+// InitializeTemplates initializes the templates needed by this controller, it must be called at controller boot time
+func InitializeTemplates(jobTemplateFileName string, cronJobTemplateFileName string) error {
+	text, err := ioutil.ReadFile(jobTemplateFileName)
 	if err != nil {
-		log.Error(err, "Error reading job template file", "filename", jobTempateFileName)
+		log.Error(err, "Error reading job template file", "filename", jobTemplateFileName)
 		return err
 	}
 	jobTemplate = template.New("Job").Funcs(template.FuncMap{
@@ -60,9 +60,9 @@ func InitializeTemplates(jobTempateFileName string, cronJobTemplateFilename stri
 		return err
 	}
 
-	text, err = ioutil.ReadFile(cronJobTemplateFilename)
+	text, err = ioutil.ReadFile(cronJobTemplateFileName)
 	if err != nil {
-		log.Error(err, "Error reading job template file", "filename", cronJobTemplateFilename)
+		log.Error(err, "Error reading job template file", "filename", cronJobTemplateFileName)
 		return err
 	}
 	cronJobTemplate = template.New("Job").Funcs(template.FuncMap{