+++ template-processors/openshift-provision/bin/processTemplates.sh
@@ -16,10 +16,9 @@
 
 set -euxo pipefail
 
-GIT_SHA=$(git rev-parse --short HEAD)
+echo "Applying resources in ${CLONED_TEMPLATE_GIT_DIR}"
 
-if [ -z "$(git status --porcelain 2>/dev/null)" ]; then
-    echo $GIT_SHA
-else
-    echo "$GIT_SHA-dirty"
-fi
+ANSIBLE_ROLES_PATH=/files/roles ansible-playbook /files/processTemplates.yml \
+  ${TEMPLATE_PROCESSOR_ARGS:-} \
+  -e template_directory="${CLONED_TEMPLATE_GIT_DIR}" \
+  -e parameter_file="${CLONED_PARAMETER_GIT_DIR}/eunomia_values_processed.yaml"